CU-8698v0xm3: Move MetaCAT tokenizers to their own package and separate implementations
