+++ tests/components/addons/meta_cat/test_meta_cat2.py
@@ -7,7 +7,7 @@ from transformers import AutoTokenizer
 
 from medcat2.components.addons.meta_cat.meta_cat import MetaCAT
 from medcat2.config.config_meta_cat import ConfigMetaCAT
-from medcat2.components.addons.meta_cat.meta_cat_tokenizers import (
+from medcat2.components.addons.meta_cat.mctokenizers.bert_tokenizer import (
     TokenizerWrapperBERT)
 from medcat2.storage.serialisers import deserialise, serialise
